data_wrangler_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data-wrangler-env"
version = "0.1.0"
description = "Data Quality & Cleaning Environment for OpenEnv — AI agents fix messy real-world datasets"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "data_wrangler_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["data_wrangler_env", "data_wrangler_env.server"]
package-dir = { "data_wrangler_env" = ".", "data_wrangler_env.server" = "server" }